1.0.0.1
1. Dodano zamianę liter w nazwie użytkownika i adresie e-mail na małe
litery w celu uniknięcia dublowania użytkowników i adresów e-mail
pisanych raz z dużej a raz z małej litery.
2. Blokada usunięcia ostatniego współpracownika (Issue #10).

+++ miniGRANT/src/main/java/pl/lukpecyn/minigrant/services/CoworkerService.java
@@ -17,6 +17,9 @@ public class CoworkerService {
 
 	@Autowired
 	JdbcTemplate jdbcTemplate;
+	
+	@Autowired 
+	BeneficiaryService beneficiaryService;
 
 	public int addCoworker(Coworker coworker) {
 		String sql ="INSERT INTO coworkers(username,beneficiary_id) VALUES(?,?)";
@@ -25,8 +28,12 @@ public class CoworkerService {
 	
 	public int delCoworker(Coworker coworker) {
 		//TODO: Chceck if Beneficiary is used
-		String sql = "DELETE FROM coworkers WHERE username=? AND beneficiary_id=?";
-		return jdbcTemplate.update(sql,coworker.getUsername(),coworker.getIdBeneficiary());
+		if (beneficiaryService.CoworkersCount(beneficiaryService.getBeneficiary(coworker.getIdBeneficiary()))>1) {
+			String sql = "DELETE FROM coworkers WHERE username=? AND beneficiary_id=?";
+			return jdbcTemplate.update(sql,coworker.getUsername(),coworker.getIdBeneficiary());
+		} else {
+			return -1;
+		}
 	}
 
 	public List<Coworker> getCoworkerList(Integer idBeneficiary) {
